refactor: transfer the files
