+++ src/features/components/pagination/Pagination.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event';
 import { MemoryRouter, Route, Routes, useLocation } from 'react-router-dom';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
-import Main from '@/features/pages/Main';
+import Main from '@/features/pages/Main/Main';
 import * as charactersHook from '@/hooks/useCharacters';
 
 import Pagination from './Pagination';
